new branch

+++ src/utils/Database.java
@@ -5,6 +5,8 @@ import java.util.Iterator;
 import java.util.Map;
 import java.util.Set;
 
+import app.InvalidFlightIDException;
+
 /**
  * Allows to create a database of airships
  * 
@@ -32,8 +34,12 @@ public class Database {
 	 * @return true if the airplane was successfully added
 	 * @return false if the airplane was not added
 	 */
-	public boolean addAirplane(Airship airplane)
+	public boolean addAirplane(Airship airplane) throws InvalidFlightIDException
 	{
+		String id = airplane.getFlightID();
+		if (id == null)
+			throw new InvalidFlightIDException();
+		
 		if(database.containsKey(airplane.getFlightID()))
 		{
 			return false;
@@ -51,8 +57,10 @@ public class Database {
 	 * @return true if the airplane was successfully removed
 	 * @return false if the specified ID was not found in the database
 	 */
-	public boolean removeAirplane(String id)
+	public boolean removeAirplane(String id) throws InvalidFlightIDException
 	{
+		if(id ==null)
+			throw new InvalidFlightIDException();
 		if(database.containsKey(id))
 		{
 			database.remove(id);
@@ -68,9 +76,11 @@ public class Database {
 	 * @return true if the airplane was successfully removed
 	 * @return false if the airplane was not found in the database
 	 */
-	public boolean removeAirplane(Airship airplane)
+	public boolean removeAirplane(Airship airplane) throws InvalidFlightIDException
 	{
 		String id = airplane.getFlightID();
+			if(id == null) 
+				throw new InvalidFlightIDException();
 		return removeAirplane(id);
 	}
 	
@@ -79,7 +89,7 @@ public class Database {
 	 * to have 0 passengers 
 	 * @return the number of airplanes that were removed from the database
 	 */
-	public int removeAirplanesWithZeroPassengers()
+	public int removeAirplanesWithZeroPassengers() throws InvalidFlightIDException
 	{
 		Set<String> idSet = database.keySet();
 		Iterator<String> iterator = idSet.iterator();
@@ -91,8 +101,12 @@ public class Database {
 			Airship airplane = database.get(iterator.next());
 			if((airplane instanceof utils.Airliner) && ((Airliner) airplane).getPassengersNumber() == 0)
 			{
+				String id = airplane.getFlightID();
+				if(id == null) 
+					throw new InvalidFlightIDException();
 				toRemove.add(airplane.getFlightID());
 				count++;
+				
 			}
 		}
 		
@@ -129,8 +143,10 @@ public class Database {
 	 * @param id the airplane id
 	 * @return true if the id is on database else returns false
 	 */
-	public boolean contains (String id)
+	public boolean contains (String id) throws InvalidFlightIDException
 	{
+		if(id == null) 
+			throw new InvalidFlightIDException();
 		return database.containsKey(id);
 	}
 }
